validateUser middleware
