+++ src/server.js
@@ -2,9 +2,9 @@ const express = require('express');
 const morgan = require('morgan');
 const cors = require('cors');
 const bcrypt = require('bcryptjs');
-const Joi = require('joi');
 const { PORT } = require('./config');
 const { addUserToDb, findUserByEmail } = require('./model/userModel');
+const { showBody, validateUser } = require('./middleware');
 
 const app = express();
 
@@ -20,20 +20,11 @@ app.use(express.json());
 app.use(showBody);
 app.use(cors());
 
-// first middlewere helper
-function showBody(req, res, next) {
-  // console.log(req.method);
-  if (req.method === 'POST') {
-    console.log('request body ===', req.body);
-  }
-  next();
-}
-
 app.get('/', (req, res) => {
   res.send('Hello World');
 });
 
-app.post('/register', async (req, res) => {
+app.post('/register', validateUser, async (req, res) => {
   // gauti vartotojo email ir password ir irasyti i users
   const { email, password } = req.body;
 
@@ -64,23 +55,9 @@ app.post('/register', async (req, res) => {
 });
 
 // POST /login - tuscias routas grazina 'bandom prisilogint'
-app.post('/login', async (req, res) => {
+app.post('/login', validateUser, async (req, res) => {
   const gautasEmail = req.body.email;
   const gautasSlaptazodis = req.body.password;
-  // validuoti gauta email ir password
-  const schema = Joi.object({
-    email: Joi.string().trim().email().lowercase().required(),
-    password: Joi.string().trim().min(5).max(10).required(),
-  });
-
-  try {
-    // abortEarly default true - rodyti tik pirma rasta klaida
-    await schema.validateAsync(req.body, { abortEarly: false });
-  } catch (error) {
-    console.log('schema.validateAsync error ===', error);
-    res.status(400).json(error.details);
-    return;
-  }
 
   // patikrinti ar yra toks email kaip gautas
   const foundUser = await findUserByEmail(gautasEmail);
